requirements.txt
opencv_python==4.1.0.25
numpy==1.16.2
setuptools==40.8.0
matplotlib==3.0.3
scipy==1.2.1
torch==1.0.0
yacs==0.1.6
graphviz==0.11.1
Pillow==6.1.0
tensorboardX==1.8
setuptools==40.8.0
